old-tests/submission/Infraware/Forms/contents/Forms/textarea_form.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_form</h3>
	</p>
	
	<hr>

	<div id="log"></div> 

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><textarea id='textarea_id'></textarea></p>
	</form>

	<script>
		
		var textarea = document.getElementById("textarea_id");
		var form = textarea.form;

		if (typeof(form) == "object") {
			test( function() {
				assert_equals(form.name, 'input_form', "form attribute is not correct.");
			});
		} else {
			test( function() {
				assert_unreached("form attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
